+++ service-api/service-typescript/workflow-control/src/api/dto/create-workflow-run-event-request.ts
@@ -1,0 +1,4 @@
+export type CreateWorkflowRunEventRequest = {
+  body: string;
+  createdBy: string;
+};